One honking good idea.
- Fixes inconsistent and off-by-one `is_in_clamp_boundary` / `is_cage_clamped`
- clone_cage_state breaks things and is superfluous
- `register_clamped_cage` should not be in charge of init-ing fdtables.
- Moves gigantic and repetitive handlers.rs file into handlers/ mod with macros
