+++ examples/namespace-grate-rs/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "namespace-grate-rs"
+name = "namespace-grate"
 version = "0.1.0"
 edition = "2024"
