+++ examples/namespace-grate-rs/src/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod clamped_lifecycle;
+pub mod ns_handlers;
+
+pub use clamped_lifecycle::*;
+pub use ns_handlers::*;
